IPRT/drv/os2: leave a hint regarding too early logging.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@75553 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Runtime/r0drv/os2/initterm-r0drv-os2.cpp
@@ -95,6 +95,6 @@ DECLHIDDEN(void) rtR0TermNative(void)
  */
 RTR0DECL(void *) RTR0Os2Virt2Flat(RTFAR16 fp)
 {
-    return (void *)KernSelToFlat((fp.sel << 16) | fp.off);
+    return (void *)KernSelToFlat(((uint32_t)fp.sel << 16) | fp.off);
 }
